requirements_video.txt
# Video/Animation Pipeline Requirements
# Core dependencies for simulation video generation and YouTube upload

# Scientific computing and visualization
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
jax>=0.4.0
jaxlib>=0.4.0

# YouTube API dependencies (optional)
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-api-python-client>=2.0.0

# Video processing
# Note: FFmpeg must be installed separately on the system
# Windows: Download from https://ffmpeg.org/download.html
# Linux: sudo apt install ffmpeg
# macOS: brew install ffmpeg

# Additional utilities
pathlib
tempfile
subprocess
json
